alx_travel_app/requirement.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0
django-environ>=0.10.0
celery>=5.3.0
mysqlclient>=2.2.0
requests>=2.31.0
